+++ wallet/src/account/operations/participation/mod.rs
@@ -12,7 +12,7 @@ pub(crate) mod event;
 pub(crate) mod voting;
 pub(crate) mod voting_power;
 
-use std::collections::{hash_map::Entry, HashMap};
+use std::collections::{hash_map::Entry, HashMap, HashSet};
 
 use iota_client::{
     api_types::plugins::participation::{
@@ -58,6 +58,17 @@ impl AccountHandle {
         log::debug!("[get_participation_overview]");
         // TODO: Could use the address endpoint in the future when https://github.com/iotaledger/inx-participation/issues/50 is done.
 
+        let mut spent_cached_outputs = self
+            .storage_manager
+            .lock()
+            .await
+            .get_cached_participation_output_status(self.read().await.index)
+            .await?;
+        let restored_spent_cached_outputs_len = spent_cached_outputs.len();
+        log::debug!(
+            "[get_participation_overview] restored_spent_cached_outputs_len: {}",
+            restored_spent_cached_outputs_len
+        );
         let outputs = self.outputs(None).await?;
         let participation_outputs = outputs
             .into_iter()
@@ -65,10 +76,13 @@ impl AccountHandle {
                 is_valid_participation_output(&output_data.output)
                 // Check that the metadata exists, because otherwise we aren't participating for anything
                     && output_data.output.features().and_then(|f| f.metadata()).is_some()
+                    // Don't add spent cached outputs, we have their data already and it can't change anymore
+                    && !spent_cached_outputs.contains_key(&output_data.output_id)
             })
             .collect::<Vec<OutputData>>();
 
         let mut events = HashMap::new();
+        let mut spent_outputs = HashSet::new();
         for output_data in participation_outputs {
             // PANIC: the filter already checks that the metadata exists.
             let metadata = output_data.output.features().and_then(|f| f.metadata()).unwrap();
@@ -88,11 +102,34 @@ impl AccountHandle {
                             entry.get_mut().push(output_data.output_id);
                         }
                     }
+                    if output_data.is_spent {
+                        spent_outputs.insert(output_data.output_id);
+                    }
                 }
             };
         }
 
         let mut participations: HashMap<ParticipationEventId, HashMap<OutputId, TrackedParticipation>> = HashMap::new();
+
+        // Add cached data
+        for (output_id, output_status_response) in &spent_cached_outputs {
+            for (event_id, participation) in &output_status_response.participations {
+                // Skip events that aren't in `event_ids` if not None
+                if let Some(event_ids) = event_ids.as_ref() {
+                    if !event_ids.contains(event_id) {
+                        continue;
+                    }
+                }
+                match participations.entry(*event_id) {
+                    Entry::Vacant(entry) => {
+                        entry.insert(HashMap::from([(*output_id, participation.clone())]));
+                    }
+                    Entry::Occupied(mut entry) => {
+                        entry.get_mut().insert(*output_id, participation.clone());
+                    }
+                }
+            }
+        }
 
         for (event_id, output_ids) in events {
             log::debug!(
@@ -104,6 +141,16 @@ impl AccountHandle {
             for output_id_chunk in output_ids.chunks(100).map(|x| x.to_vec()) {
                 let mut tasks = Vec::new();
                 for output_id in output_id_chunk {
+                    // Skip if participations already contains this output id with participation for this event
+                    if let Some(p) = participations.get(&event_id) {
+                        if p.contains_key(&output_id) {
+                            log::debug!(
+                                "[get_participation_overview] skip requesting already known {output_id} for event {event_id}",
+                            );
+                            continue;
+                        }
+                    }
+
                     let event_client = event_client.clone();
                     tasks.push(async move {
                         task::spawn(async move { (event_client.output_status(&output_id).await, output_id) }).await
@@ -114,6 +161,22 @@ impl AccountHandle {
                 for (result, output_id) in results {
                     match result {
                         Ok(status) => {
+                            // Cache data for spent outputs
+                            if spent_outputs.contains(&output_id) {
+                                match spent_cached_outputs.entry(output_id) {
+                                    Entry::Vacant(entry) => {
+                                        entry.insert(status.clone());
+                                    }
+                                    Entry::Occupied(mut entry) => {
+                                        let output_status_response = entry.get_mut();
+                                        for (event_id, participation) in &status.participations {
+                                            output_status_response
+                                                .participations
+                                                .insert(*event_id, participation.clone());
+                                        }
+                                    }
+                                }
+                            }
                             for (event_id, participation) in status.participations {
                                 // Skip events that aren't in `event_ids` if not None
                                 if let Some(event_ids) = event_ids.as_ref() {
@@ -138,6 +201,19 @@ impl AccountHandle {
             }
         }
 
+        log::debug!(
+            "[get_participation_overview] new spent_cached_outputs: {}",
+            spent_cached_outputs.len()
+        );
+        // Only store updated data if new outputs got added
+        if spent_cached_outputs.len() > restored_spent_cached_outputs_len {
+            self.storage_manager
+                .lock()
+                .await
+                .set_cached_participation_output_status(self.read().await.index, spent_cached_outputs)
+                .await?;
+        }
+
         Ok(AccountParticipationOverview { participations })
     }
 
@@ -145,6 +221,7 @@ impl AccountHandle {
     ///
     /// If multiple outputs with this tag exist, the one with the largest amount will be returned.
     pub async fn get_voting_output(&self) -> Result<Option<OutputData>> {
+        log::debug!("[get_voting_output]");
         Ok(self
             .unspent_outputs(None)
             .await?
@@ -157,6 +234,7 @@ impl AccountHandle {
     /// Gets client for an event.
     /// If event isn't found, the client from the account will be returned.
     pub(crate) async fn get_client_for_event(&self, id: &ParticipationEventId) -> crate::Result<Client> {
+        log::debug!("[get_client_for_event]");
         let account_index = self.read().await.index;
         let events = self
             .storage_manager
@@ -183,6 +261,7 @@ impl AccountHandle {
         &self,
         participations: &mut Participations,
     ) -> crate::Result<()> {
+        log::debug!("[remove_ended_participation_events]");
         let latest_milestone_index = self.client().get_info().await?.node_info.status.latest_milestone.index;
 
         let account_index = self.read().await.index;