+++ wallet/src/storage/constants.rs
@@ -27,3 +27,5 @@ pub(crate) const DATABASE_SCHEMA_VERSION_KEY: &str = "database-schema-version";
 
 #[cfg(feature = "participation")]
 pub(crate) const PARTICIPATION_EVENTS: &str = "participation-events";
+#[cfg(feature = "participation")]
+pub(crate) const PARTICIPATION_CACHED_OUTPUTS: &str = "participation-cached-outputs";